refactor(tui): extract pure key handling into tui/keys.rs
Move AppState::handle_key and its per-screen handle_key_* methods into a
child keys module as a second inherent impl block. The handlers reach the
state methods that stay in mod.rs via the descendant-sees-ancestor-privates
rule; only the public handle_key entry is called from run_loop. Pure
refactor, no behaviour change. Completes the tui.rs split for #73.
